% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_search.R
\name{check_tables}
\alias{check_tables}
\title{Verificar parâmetros disponíveis}
\usage{
check_tables(dir)
}
\arguments{
\item{dir}{Pasta onde base de tags está armazenada.}
}
\value{
Um vetor de caracteres
}
\description{
Verificar parâmetros disponíveis
}
\examples{
tags_dir <- create_sqlite_tags()
check_tables(tags_dir)
}
