% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_tag_creation.R
\name{create_sqlite_tags}
\alias{create_sqlite_tags}
\title{Cria arquivo sqlite com tags de consulta na API do Tabnet ANS}
\usage{
create_sqlite_tags()
}
\value{
Um vetor de caracteres com o caminho do arquivo.
}
\description{
Cria arquivo sqlite com tags de consulta na API do Tabnet ANS
}
\examples{
tags_dir <- create_sqlite_tags()
}
